hook regarding returnToLastScene, move hooks into source files
